requirements.txt
# Video Engine (existing)
moviepy>=2.0
Pillow>=10.0
librosa>=0.10
edge-tts>=6.1
numpy>=1.24

# PocketBase Worker
httpx>=0.27
